+++ chalicelib/payments.py
@@ -219,8 +219,8 @@ def check_customer_account_status(customer):
 
     invoice = stripe.Invoice.upcoming(customer=customer.id)
 
-    # We're going to treat polyverse accounts as paid, even though no credit card on file
-    if customer['email'].endswith("@polyverse.io") or customer['email'].endswith("@polytest.ai"):
+    # We're going to treat polyverse accounts as paid, even if no credit card on file
+    if customer['email'].endswith(("@polyverse.io", "@polytest.ai", "@polyverse.com")):
         return True, "paid"
 
     # Check if the customer has a default payment method (or manually entered payment)